+++ src/main/kotlin/icu/windea/pls/lang/navigation/ParadoxDefinitionChooseByNameContributor.kt
@@ -27,7 +27,7 @@ class ParadoxDefinitionChooseByNameContributor : ChooseByNameContributorEx {
     override fun processNames(processor: Processor<in String>, scope: GlobalSearchScope, filter: IdFilter?) {
         if (!isEnabled()) return
         val project = scope.project ?: getCurrentProject() ?: return
-        val gameType = ParadoxAnalysisManager.getInferredCurrentGameType(project)
+        val gameType = ParadoxAnalysisManager.getSelectedGameType(project)
         val selector = ParadoxDefinitionSearch.selector(project).withSearchScope(scope).withGameType(gameType)
         ParadoxDefinitionSearch.search(null, null, selector).process p@{
             val name = it.name
@@ -40,7 +40,7 @@ class ParadoxDefinitionChooseByNameContributor : ChooseByNameContributorEx {
         val name = PlsChooseByNameUtil.getAdjustedName(name, parameters) // adjust name if necessary
         val project = parameters.project
         val scope = GlobalSearchScopeUtil.toGlobalSearchScope(parameters.searchScope, project)
-        val gameType = ParadoxAnalysisManager.getInferredCurrentGameType(project)
+        val gameType = ParadoxAnalysisManager.getSelectedGameType(project)
         val selector = ParadoxDefinitionSearch.selector(project).withSearchScope(scope).withGameType(gameType)
         ParadoxDefinitionSearch.search(name, null, selector).process p@{
             val element = it.element ?: return@p true